restore Home formatting by moving the Home header to a new component
